tests/wpt/tests/html/canvas/offscreen/transformations/2d.transformation.transform.nonfinite.html

<!DOCTYPE html>
<!-- DO NOT EDIT! This test has been generated by /html/canvas/tools/gentest.py. -->
<meta charset="UTF-8">
<title>OffscreenCanvas test: 2d.transformation.transform.nonfinite</title>
<script src="/resources/testharness.js"></script>
<script src="/resources/testharnessreport.js"></script>
<script src="/html/canvas/resources/canvas-tests.js"></script>

<h1>2d.transformation.transform.nonfinite</h1>
<p class="desc">transform() with Infinity/NaN is ignored</p>


<script>
var t = async_test("transform() with Infinity/NaN is ignored");
var t_pass = t.done.bind(t);
var t_fail = t.step_func(function(reason) {
    throw reason;
});
t.step(function() {

  var canvas = new OffscreenCanvas(100, 50);
  var ctx = canvas.getContext('2d');

  ctx.fillStyle = '#f00';
  ctx.fillRect(0, 0, 100, 50);

  ctx.translate(100, 10);
  ctx.transform(Infinity, 0, 0, 0, 0, 0);
  ctx.transform(-Infinity, 0, 0, 0, 0, 0);
  ctx.transform(NaN, 0, 0, 0, 0, 0);
  ctx.transform(0, Infinity, 0, 0, 0, 0);
  ctx.transform(0, -Infinity, 0, 0, 0, 0);
  ctx.transform(0, NaN, 0, 0, 0, 0);
  ctx.transform(0, 0, Infinity, 0, 0, 0);
  ctx.transform(0, 0, -Infinity, 0, 0, 0);
  ctx.transform(0, 0, NaN, 0, 0, 0);
  ctx.transform(0, 0, 0, Infinity, 0, 0);
  ctx.transform(0, 0, 0, -Infinity, 0, 0);
  ctx.transform(0, 0, 0, NaN, 0, 0);
  ctx.transform(0, 0, 0, 0, Infinity, 0);
  ctx.transform(0, 0, 0, 0, -Infinity, 0);
  ctx.transform(0, 0, 0, 0, NaN, 0);
  ctx.transform(0, 0, 0, 0, 0, Infinity);
  ctx.transform(0, 0, 0, 0, 0, -Infinity);
  ctx.transform(0, 0, 0, 0, 0, NaN);
  ctx.transform(Infinity, Infinity, 0, 0, 0, 0);
  ctx.transform(Infinity, Infinity, Infinity, 0, 0, 0);
  ctx.transform(Infinity, Infinity, Infinity, Infinity, 0, 0);
  ctx.transform(Infinity, Infinity, Infinity, Infinity, Infinity, 0);
  ctx.transform(Infinity, Infinity, Infinity, Infinity, Infinity, Infinity);
  ctx.transform(Infinity, Infinity, Infinity, Infinity, 0, Infinity);
  ctx.transform(Infinity, Infinity, Infinity, 0, Infinity, 0);
  ctx.transform(Infinity, Infinity, Infinity, 0, Infinity, Infinity);
  ctx.transform(Infinity, Infinity, Infinity, 0, 0, Infinity);
  ctx.transform(Infinity, Infinity, 0, Infinity, 0, 0);
  ctx.transform(Infinity, Infinity, 0, Infinity, Infinity, 0);
  ctx.transform(Infinity, Infinity, 0, Infinity, Infinity, Infinity);
  ctx.transform(Infinity, Infinity, 0, Infinity, 0, Infinity);
  ctx.transform(Infinity, Infinity, 0, 0, Infinity, 0);
  ctx.transform(Infinity, Infinity, 0, 0, Infinity, Infinity);
  ctx.transform(Infinity, Infinity, 0, 0, 0, Infinity);
  ctx.transform(Infinity, 0, Infinity, 0, 0, 0);
  ctx.transform(Infinity, 0, Infinity, Infinity, 0, 0);
  ctx.transform(Infinity, 0, Infinity, Infinity, Infinity, 0);
  ctx.transform(Infinity, 0, Infinity, Infinity, Infinity, Infinity);
  ctx.transform(Infinity, 0, Infinity, Infinity, 0, Infinity);
  ctx.transform(Infinity, 0, Infinity, 0, Infinity, 0);
  ctx.transform(Infinity, 0, Infinity, 0, Infinity, Infinity);
  ctx.transform(Infinity, 0, Infinity, 0, 0, Infinity);
  ctx.transform(Infinity, 0, 0, Infinity, 0, 0);
  ctx.transform(Infinity, 0, 0, Infinity, Infinity, 0);
  ctx.transform(Infinity, 0, 0, Infinity, Infinity, Infinity);
  ctx.transform(Infinity, 0, 0, Infinity, 0, Infinity);
  ctx.transform(Infinity, 0, 0, 0, Infinity, 0);
  ctx.transform(Infinity, 0, 0, 0, Infinity, Infinity);
  ctx.transform(Infinity, 0, 0, 0, 0, Infinity);
  ctx.transform(0, Infinity, Infinity, 0, 0, 0);
  ctx.transform(0, Infinity, Infinity, Infinity, 0, 0);
  ctx.transform(0, Infinity, Infinity, Infinity, Infinity, 0);
  ctx.transform(0, Infinity, Infinity, Infinity, Infinity, Infinity);
  ctx.transform(0, Infinity, Infinity, Infinity, 0, Infinity);
  ctx.transform(0, Infinity, Infinity, 0, Infinity, 0);
  ctx.transform(0, Infinity, Infinity, 0, Infinity, Infinity);
  ctx.transform(0, Infinity, Infinity, 0, 0, Infinity);
  ctx.transform(0, Infinity, 0, Infinity, 0, 0);
  ctx.transform(0, Infinity, 0, Infinity, Infinity, 0);
  ctx.transform(0, Infinity, 0, Infinity, Infinity, Infinity);
  ctx.transform(0, Infinity, 0, Infinity, 0, Infinity);
  ctx.transform(0, Infinity, 0, 0, Infinity, 0);
  ctx.transform(0, Infinity, 0, 0, Infinity, Infinity);
  ctx.transform(0, Infinity, 0, 0, 0, Infinity);
  ctx.transform(0, 0, Infinity, Infinity, 0, 0);
  ctx.transform(0, 0, Infinity, Infinity, Infinity, 0);
  ctx.transform(0, 0, Infinity, Infinity, Infinity, Infinity);
  ctx.transform(0, 0, Infinity, Infinity, 0, Infinity);
  ctx.transform(0, 0, Infinity, 0, Infinity, 0);
  ctx.transform(0, 0, Infinity, 0, Infinity, Infinity);
  ctx.transform(0, 0, Infinity, 0, 0, Infinity);
  ctx.transform(0, 0, 0, Infinity, Infinity, 0);
  ctx.transform(0, 0, 0, Infinity, Infinity, Infinity);
  ctx.transform(0, 0, 0, Infinity, 0, Infinity);
  ctx.transform(0, 0, 0, 0, Infinity, Infinity);

  ctx.fillStyle = '#0f0';
  ctx.fillRect(-100, -10, 100, 50);

  _assertPixel(canvas, 50,25, 0,255,0,255);
  t.done();

});
</script>
