LayoutTests/imported/w3c/web-platform-tests/html/canvas/offscreen/transformations/2d.transformation.transform.nonfinite-expected.txt

2d.transformation.transform.nonfinite

transform() with Infinity/NaN is ignored


PASS transform() with Infinity/NaN is ignored 
